Adding product stats endpoint
    /productStats/:channel/:id

returns something that looks like this

{
  "TotalRevenue": 0,
  "TotalOrders": 0,
  "AverageNumberPerOrder": 0,
  "TotalInCarts": 18,
  "ProductConversionRate": 0,
  "AverageTotalRevenue": 0,
  "AverageTotalOrders": 0,
  "AverageAverageNumberPerOrder": 0,
  "AverageTotalInCarts": 9,
  "AverageProductConversionRate": 0
}

+++ intelligence/eggcrate/src/responses/productFunnelResponse.go
@@ -6,8 +6,8 @@ type ProductFunnelResponse struct {
 	CartClicks        int
 	CheckoutClicks    int
 	Purchases         int
-	SearchToPdp       float32
-	PdpToCart         float32
-	CartToCheckout    float32
-	CheckoutPurchased float32
+	SearchToPdp       float64
+	PdpToCart         float64
+	CartToCheckout    float64
+	CheckoutPurchased float64
 }
